feat: implement a majority judgment deliberator
This should work (at least, the test-suite passes).
This is our first java library, and we're newbies,
so be kind and do patch what's gouging your eyes.

This implementation is score-based, for performance.
There are many other ways of implementing MJ.

There is no support (yet) for a default grade,
but you can do it yourself while building the tally.

Special thanks to @plguhur for the assistance,
and the whole MieuxVoter operational team.

/spend 36h